move mmap structs to mmap.h and make c++ compatible
